Move temporaryDirectory() out of Filesystem and into a global val
I'd considered making a TEMPORARY FileSystem rooted at this directory,
but it's annoying to interop with Java I/O or user-visible paths. This
is smaller and simpler.
