+++ okio-files/src/commonMain/kotlin/okio/Platform.kt
@@ -16,4 +16,7 @@
 package okio
 
 internal expect val PLATFORM_FILESYSTEM: Filesystem
+
+internal expect val PLATFORM_TEMPORARY_DIRECTORY: Path
+
 internal expect val DIRECTORY_SEPARATOR: String
