+++ okio-files/src/jvmMain/kotlin/okio/JvmSystemFilesystem.kt
@@ -15,7 +15,6 @@
  */
 package okio
 
-import okio.Path.Companion.toPath
 import java.nio.file.Files
 import java.nio.file.StandardCopyOption.ATOMIC_MOVE
 import java.nio.file.StandardCopyOption.REPLACE_EXISTING
@@ -61,8 +60,6 @@ object JvmSystemFilesystem : Filesystem() {
   ) {
     commonCopy(source, target)
   }
-
-  override fun temporaryDirectory() = System.getProperty("java.io.tmpdir").toPath()
 
   override fun delete(path: Path) {
     val deleted = path.toFile().delete()
